+++ packages/blocks/src/root-block/edgeless/components/block-portal/note/edgeless-note.ts
@@ -1,11 +1,6 @@
 import '../../note-slicer/index.js';
 
-import {
-  type BlockElement,
-  ShadowlessElement,
-  WithDisposable,
-} from '@blocksuite/lit';
-import type { BlockModel } from '@blocksuite/store';
+import { ShadowlessElement, WithDisposable } from '@blocksuite/lit';
 import { css, html, nothing } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
@@ -17,7 +12,10 @@ import { MoreIndicatorIcon } from '../../../../../_common/icons/edgeless.js';
 import { NoteDisplayMode } from '../../../../../_common/types.js';
 import { almostEqual, clamp } from '../../../../../_common/utils/math.js';
 import { matchFlavours } from '../../../../../_common/utils/model.js';
-import { buildPath } from '../../../../../_common/utils/query.js';
+import {
+  buildPath,
+  findClosestTextBlock,
+} from '../../../../../_common/utils/query.js';
 import { type NoteBlockModel } from '../../../../../note-block/note-model.js';
 import { Bound, StrokeStyle } from '../../../../../surface-block/index.js';
 import type { SurfaceBlockComponent } from '../../../../../surface-block/surface-block.js';
@@ -180,44 +178,6 @@ export class EdgelessBlockPortalNote extends EdgelessPortalBase<NoteBlockModel> 
     }
   }
 
-  private _findClosestTextBlock(
-    blocks: BlockModel[],
-    x: number,
-    y: number
-  ): BlockElement | null {
-    for (let i = 0; i < blocks.length; i++) {
-      const block = blocks[i];
-      const blockElement = this.edgeless.host.view.viewFromPath(
-        'block',
-        buildPath(block)
-      );
-      if (!blockElement) continue;
-      const blockRect = blockElement.getBoundingClientRect();
-      if (y >= blockRect.top && y <= blockRect.bottom) {
-        const childrenTextBlocks = block.children.filter(model =>
-          matchFlavours(model, [
-            'affine:paragraph',
-            'affine:code',
-            'affine:list',
-          ])
-        );
-        // If no children, return the block
-        // If has children, find the closest one, if not found, return the block
-        if (!childrenTextBlocks.length) {
-          return blockElement;
-        } else {
-          const childBlock = this._findClosestTextBlock(
-            childrenTextBlocks,
-            x,
-            y
-          );
-          return childBlock ? childBlock : blockElement;
-        }
-      }
-    }
-    return null;
-  }
-
   private _handleClickAtBackground(e: MouseEvent) {
     e.stopPropagation();
     if (!this._affineNote || !this._editing) return;
@@ -234,12 +194,17 @@ export class EdgelessBlockPortalNote extends EdgelessPortalBase<NoteBlockModel> 
       matchFlavours(model, ['affine:paragraph', 'affine:code', 'affine:list'])
     );
 
-    const closestBlockElement = this._findClosestBlock(textBlocks, x, y);
-    if (closestBlockElement) {
-      const textPath = buildPath(closestBlockElement.model);
+    const closestTextBlockElement = findClosestTextBlock(
+      textBlocks,
+      this.edgeless.host.view,
+      x,
+      y
+    );
+    if (closestTextBlockElement) {
+      const textPath = buildPath(closestTextBlockElement.model);
       this.edgeless.host.selection.clear(['text']);
-      let index = closestBlockElement.model.text?.length ?? 0;
-      const rect = closestBlockElement.getBoundingClientRect();
+      let index = closestTextBlockElement.model.text?.length ?? 0;
+      const rect = closestTextBlockElement.getBoundingClientRect();
       const originX = e.x;
       const originY = e.y;
       if (originX < rect.left || originY < rect.top) {
